Template and wrapper subcommands
Added subcommands for template and wrapper generation
Running actions is now under a subcommand
Combined verbose and debug options
Updating some file names
