+++ src/bdast/__init__.py
@@ -1,3 +1,3 @@
-from .bdast import main
+from .cli import main
 
 __all__ = ["main"]
