+++ src/bdast/__main__.py
@@ -1,4 +1,4 @@
-from .bdast import main
+from .cli import main
 
 if __name__ == "__main__":
     main()
